Complete the 10.3.3 documentation, ADR amendment and roadmap tick
Milestone 8 of roadmap 10.3.3:

- ADR-010 moves to Accepted with dated adjacent amendments that preserve
  the original "absolute-path include_str! is rejected" wording verbatim
  while recording the deferred-path measurement that scopes the rejection
  to token-stream literals (D0a), the directory-scan correction to the
  Option B per-file claim (D2), and the constraint-5 rewrite (D5). The
  Binary-size table row is corrected.
- The design document's §2.7.6.6 is rewritten from "foot-gun and two
  candidate mechanisms" to "here is what ships and why" (the anonymous
  `include_bytes!` binding, the D3 relative embedding, the tested
  `build.rs` recipe closing the addition gap).
- The v0.6.0 migration guide deletes the stale "feature-file edits do not
  trigger a rebuild" caveat, adds the manifest-relative feature-path
  breaking-change section (with the JUnit `classname` discontinuity,
  the outside-manifest fallback and the workspace-uniqueness loss), the
  D4 unrelatable-root `compile_error!` announcement, and the hermetic
  build-system ("undeclared dependency") paragraph, plus checklist
  entries for the three.
- The users-guide caveats are replaced, the testing-strategy and
  developers-guide documents gain the new test classes and house
  conventions, repository-layout names the new fixture paths, the
  changelog records the two user-visible changes, and the roadmap ticks
  10.3.3 and queues three follow-up items (10.3.4–10.3.6).

Supporting changes: the trybuild dep-info assertion moves to a
`trybuild_macros/tracking.rs` sibling module (the main file crossed the
400-line cap) and now normalizes both the needle and the depot — a
Windows-only failure — and the recipe's exact `rerun-if-changed` line is
pinned by the extractor test because Cargo treats a missing watch path
as perpetually dirty.

All gates green: check-fmt, lint (incl. Whitaker), test (1726/1726),
markdown/spelling/nixie, and publish-check.
