+++ docs/adr-010-feature-file-change-detection.md
@@ -2,11 +2,16 @@
 
 ## Status
 
-Proposed
+Accepted (2026-08-17): the implementing ExecPlan (roadmap item 10.3.3) chose a
+macro-emitted `include_bytes!` tracking binding at item scope (one anonymous
+`const` per bound `.feature` file, path deferred through
+`concat!(env!("CARGO_MANIFEST_DIR"), "/", <relative literal>)`), plus a tested
+`build.rs` recipe for the file-addition gap. The amendments below record the
+measurements that corrected the original rejection rationale.
 
 ## Date
 
-2026-06-10
+2026-06-10; amended 2026-08-17.
 
 ## Context and problem statement
 
@@ -91,6 +96,24 @@ is rejected because:
 
 This option variant is therefore not a valid implementation of Option A.
 
+**Amendment adopted 2026-08-17: the rejection above applies to a proc macro
+that writes an absolute *string literal* into the token stream — not to the
+`concat!(env!("CARGO_MANIFEST_DIR"), "/", REL)` deferred form.** The variant
+measured in the implementing ExecPlan (transcripts A and B) emits
+`env!("CARGO_MANIFEST_DIR")` and lets rustc construct the path at expansion
+time, so no absolute path appears in the emitted token stream, and the unused
+anonymous `const` is elided: control binaries and binaries carrying the binding
+showed the same absolute-path presence in dev profile (from debug info; release
+binaries clean in both) and no feature text in either, so the mechanism adds
+neither the path nor the file contents to a compiled binary. The `rlib` retains
+source paths and constant values in metadata regardless, so the ADR's
+binary-size and path claims are scoped to binary and test targets. The shipping
+form therefore uses `include_bytes!` (which, unlike `include_str!`, does not
+hard-error on a non-UTF-8 `.feature` file yet registers dep-info identically)
+and an anonymous `const _` — naming the constant would make the file contents
+reachable and retained, the cost this ADR feared. See the ExecPlan Decision
+D0/D0a.
+
 ### Option B: build-script helper emitting `cargo::rerun-if-changed`
 
 Provide an `rstest-bdd-build` helper crate (or a function in `rstest-bdd`) that
@@ -131,6 +154,16 @@ Cons:
   even when absent (Cargo ignores directives for non-existent paths), and scan
   recursively.
 
+**Amendment adopted 2026-08-17: the allow-list concern above is over-cautious
+for the directory form.** Cargo scans a `rerun-if-changed` **directory**
+recursively (rust-lang/cargo#8973), so a single directory line covers newly
+added files and subdirectories without per-file emission — a hand-maintained
+per-file list is the more fragile option, because it can silently omit a new
+subdirectory. The ExecPlan ships exactly that one-line recipe as *tested living
+documentation*: a test extracts it from the users' guide, writes it into a
+fixture crate, and proves a newly added `.feature` file is picked up by the next
+`cargo test` (Decision D2).
+
 ### Option C: `proc_macro::tracked_path` (unstable)
 
 Use the unstable `proc_macro::tracked_path::path()` API, which is the primitive
@@ -166,7 +199,7 @@ foot-gun. Recorded here to keep the analysis complete; addressed separately in
 | Axis                | A (relative `include_str!`) | B (build script)      | C (`tracked_path`) | D (OUT_DIR cache) |
 | ------------------- | --------------------------- | --------------------- | ------------------ | ----------------- |
 | Consumer-invisible  | High                        | Low                   | High               | Low               |
-| Binary-size cost    | Medium                      | None                  | None               | None              |
+| Binary-size cost    | Low                         | None                  | None               | None              |
 | Absolute-path risk  | None (relative only)        | None                  | None               | None              |
 | `scenarios!` fit    | Medium (one per file)       | High (directory scan) | High               | None              |
 | Reproducible builds | High                        | High                  | High               | N/A               |
@@ -201,8 +234,11 @@ Binding constraints for the implementing ExecPlan:
    reliable relative path that rustc resolves correctly. This is the path of
    zero consumer friction.
 5. **Option C (`tracked_path`)** is recorded as the right long-term answer.
-   Usable behind a `nightly` feature gate; stabilization should be monitored
-   and adopted when available.
+   Not adoptable behind a `nightly` feature gate while the workspace's own
+   gates run `--all-features` on a stable toolchain (which would enable it and
+   break them); revisit on stabilization. It also buys nothing over
+   `include_bytes!` here — both register paths in dep-info, and neither closes
+   the directory-addition gap.
 6. **Option D (OUT_DIR cache)** is out of scope for this ADR. It addresses
    compile performance, not invalidation correctness; `§3.2.2` tracks it
    separately.
