+++ docs/testing-strategy.md
@@ -124,3 +124,52 @@ Isolation is therefore per-thread; any test that reads from shared state must
 call the corresponding reset helper before running its scenario. Tests that
 mutate shared state must be annotated with `#[serial]` to prevent interleaving
 with other tests on the same thread pool.
+
+## Cargo-spawning fixture-crate tests
+
+Roadmap 10.3.3 introduced a distinct class of test: a fixture crate is copied
+into the shared workspace `target/`, its manifest's relative path dependencies
+are rewritten to absolute paths, and a nested `cargo` invocation is driven
+against it with a byte-identical child environment. The strategy guarantees
+that apply:
+
+- The checked-in fixture is never mutated; all edits happen in the scratch
+  copy, so a killed test is recovered by deleting the scratch directory.
+- The child environment differs from the parent's only where cross-talk or
+  stalls would result (`CARGO_MAKEFLAGS`, `CARGO_PKG_*` stripped;
+  `LLVM_PROFILE_FILE` redirected; `CARGO_TARGET_DIR` inherited or defaulted to
+  the workspace target).
+- Every nested `cargo` invocation runs under the harness's own wall-clock
+  bound, and the binary is serialized against other cargo-spawning tests
+  through the `cargo-spawning` nextest test-group.
+- The regression suite's dep-info assertion is a direct filesystem check
+  (rustc's `.d` file), deliberately _not_ mediated by the macros under test, so
+  a macro bug cannot mask the regression it is meant to prove.
+
+See `crates/rstest-bdd/tests/feature_rebuild_invalidation/` for the worked
+harness and `docs/developers-guide.md` for the conventions.
+
+## Tested living documentation
+
+A second new class of test executes fenced examples extracted from user-facing
+Markdown (roadmap 10.3.3, Milestone 7, modelled on
+[`netsuke`](https://github.com/leynos/netsuke)). Markers
+(`<!-- tested-example: id -->`) key each executable example; enforcement is
+regional so the documentation cannot quietly acquire an untested example inside
+an enforced section. The recipe in the users-guide's rebuild invalidation
+section is the first such example and is executed end-to-end: it is written
+into a fixture crate's `build.rs` and a behavioural test proves a newly added
+`.feature` file is run. The rule of thumb: prose that must not rot when
+executed — recipes, configs, commands — should be marked and consumed by a test
+rather than duplicated.
+
+## Assertion posture
+
+This repository adopted `googletest` and `pretty_assertions` in roadmap 10.3.3
+(ExecPlan Decision D1, previously unused anywhere in the workspace). Use
+`assert_that!` / `expect_that!` and matchers where the assertion expresses a
+property (`contains_substring`, `eq`, `is_true`, `len`, `each`), and
+`pretty_assertions` where a structural-equality diff carries the value.
+`expect_that!` (which aggregates several failures into one report) requires the
+`#[gtest]` attribute; inside `#[scenario]`-generated bodies there is no
+`#[gtest]` context, so step functions use the panic-mode `assert_that!`.
