+++ docs/users-guide.md
@@ -1672,8 +1672,7 @@ binding name is part of the contract.
   test][gpui-migration] subsection (inside "Adopt GPUI harness configuration")
   walks readers through moving an existing scenario to the playbook.
 - Design-document §2.7.6.6 documents the feature-file rebuild-invalidation
-  foot-gun (`.feature`-only edits do not trigger a rebuild until roadmap item
-  10.3.3 lands).
+  mechanism (`.feature`-only edits recompile the scenario binary).
 - Design-document §2.7.6.7 documents the full cargo test versus nextest matrix
   for `#[serial]` and thread-local state.
 
@@ -1815,9 +1814,9 @@ published `gpui 0.2.2` — the audience migrating real suites — should adapt t
 using the vendored-to-published mapping table above (under "Durable handles
 versus visual context").
 
-Editing only a `.feature` file does not trigger a rebuild (see design-document
-§2.7.6.6), so touch a binding `.rs` file (or run `cargo clean -p <crate>`)
-after changing feature text; otherwise a stale build can mask the change.
+Editing only a `.feature` file triggers a rebuild of the scenario binary (see
+[Feature file rebuild invalidation](#feature-file-rebuild-invalidation)), so no
+`touch` or `cargo clean` workaround is needed after changing feature text.
 
 ### Test-runner parallelism and scenario state
 
@@ -3235,21 +3234,21 @@ three amigos in the specification process.
 ### Feature file rebuild invalidation
 
 Since v0.6.0, `#[scenario]` and `scenarios!` register every bound `.feature`
-file as a Cargo rebuild dependency: the macro emits an `include_bytes!`
-binding whose path is built from `CARGO_MANIFEST_DIR`, so rustc records the
-file in dep-info and Cargo recompiles the scenario binary whenever the file
-changes. Editing only a `.feature` file therefore recompiles the crate, and
-the tests reflect the new text immediately.
+file as a Cargo rebuild dependency: the macro emits an `include_bytes!` binding
+whose path is built from `CARGO_MANIFEST_DIR`, so rustc records the file in
+dep-info and Cargo recompiles the scenario binary whenever the file changes.
+Editing only a `.feature` file therefore recompiles the crate, and the tests
+reflect the new text immediately.
 
 `#[scenario]` tracks its single bound file. `scenarios!` tracks every
 `.feature` file it discovers in the bound directory — including files whose
 scenarios are excluded by a `tags =` filter.
 
-Per-file tracking covers *edits* to files that existed when the macro ran.
-It cannot see a file that is *added* afterwards, because nothing referenced
-a file that did not exist at expansion time. To close that gap, add a
-`build.rs` to your crate that tells Cargo to rerun the build script
-whenever the bound directory changes:
+Per-file tracking covers *edits* to files that existed when the macro ran. It
+cannot see a file that is *added* afterwards, because nothing referenced a file
+that did not exist at expansion time. To close that gap, add a `build.rs` to
+your crate that tells Cargo to rerun the build script whenever the bound
+directory changes:
 
 <!-- tested-example: scenarios-build-script -->
 
@@ -3260,18 +3259,17 @@ fn main() {
 ```
 
 Cargo scans a `rerun-if-changed` **directory** recursively, so a single
-directory line covers new files and subdirectories; a hand-maintained
-per-file list can silently omit a new subdirectory, which is the failure
-mode this mechanism exists to prevent. The snippet above is executed by
-this repository's test suite (it is extracted from this document and run
-against a scratch crate, see the developers' guide), so it cannot rot.
+directory line covers new files and subdirectories; a hand-maintained per-file
+list can silently omit a new subdirectory, which is the failure mode this
+mechanism exists to prevent. The snippet above is executed by this repository's
+test suite (it is extracted from this document and run against a scratch crate,
+see the developers' guide), so it cannot rot.
 
 #### Cost
 
-Per-file tracking makes a previously free operation cost something:
-editing one `.feature` file now rebuilds the whole test binary containing
-the `scenarios!` invocation that binds it, which re-parses every feature
-file in the bound directory. Measured at roughly 5.4 s for 100 files with
-five scenarios each on a development host. For large feature directories,
-split them across several test binaries with disjoint `scenarios!`
-invocations.
+Per-file tracking makes a previously free operation cost something: editing one
+`.feature` file now rebuilds the whole test binary containing the `scenarios!`
+invocation that binds it, which re-parses every feature file in the bound
+directory. Measured at roughly 5.4 s for 100 files with five scenarios each on
+a development host. For large feature directories, split them across several
+test binaries with disjoint `scenarios!` invocations.
